MOTECH-1379: Improve Commcare Performance
This commit improves commcare schema saving time. Instead of using one-to-many
relations in Commcare module we insert serialized objects into MOTECH database.
Additionally now we update Commcare schema only if needed.

The saving time is much better. More details in attached document on JIRA.

Unfortunately @Ignore doesn't work correctly and before fixing this problem,
we need also @NotPersistent annotation.

+++ commcare/src/main/java/org/motechproject/commcare/events/AppSchemaChangeEventHandler.java
@@ -34,12 +34,15 @@ public class AppSchemaChangeEventHandler {
     @MotechListener(subjects = SCHEMA_CHANGE_EVENT)
     public void schemaChange(MotechEvent event) {
         List<CommcareApplicationJson> applications = appStructureService.getAllApplications();
-        commcareApplicationDataService.deleteAll();
 
-        for (CommcareApplicationJson app : applications) {
-            commcareApplicationDataService.create(app);
+        if (!applications.equals(commcareApplicationDataService.retrieveAll())) {
+            commcareApplicationDataService.deleteAll();
+
+            for (CommcareApplicationJson app : applications) {
+                commcareApplicationDataService.create(app);
+            }
+
+            commcareTasksNotifier.updateTasksInfo();
         }
-
-        commcareTasksNotifier.updateTasksInfo();
     }
 }
